requirements.txt
fastapi
uvicorn
google-genai
google-play-scraper
requests